requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-genai==0.3.0
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.1.0
aiofiles==23.2.1
pydantic==2.5.0
jinja2==3.1.2
toml==0.10.2
requests==2.31.0
